+++ algo/source/functions.h
@@ -24,8 +24,8 @@ void swapTwoItems(int * solution, int size);
 void loadInitialSolution(int *solution, int size);
 
 /* Send a solution to the server to receive the solution's fitness value*/
-int askFitness(const int *solution, int size);
+double askFitness(const int *solution, int size);
 
-int chooseBetterSolution(const int * solution, int size, int bestFitness, int nbEval, int * betterSolution);
+double chooseBetterSolution(const int * solution, int size, double bestFitness, int nbEval, int * betterSolution);
 
 #endif
