+++ algo/source/main.c
@@ -16,21 +16,21 @@ int main(int argc, char *argv[]) {
 
     loadInitialSolution(solution, size);
 
-    int fitness = askFitness(solution, size);
-    printf("Fitness value for solution initial = %d\n", fitness);
+    double fitness = askFitness(solution, size);
+    printf("Fitness value for solution initial = %f\n", fitness);
 
     for (int i = 0; i < nbEvalMax; ++i) {
         int potentialBetterSolution[size];
-        int  potentialBetterFitness = chooseBetterSolution(solution,size,fitness,nbEvalMax, potentialBetterSolution);
+        double  potentialBetterFitness = chooseBetterSolution(solution,size,fitness,nbEvalMax, potentialBetterSolution);
 
         if(potentialBetterFitness < fitness){
             fitness = potentialBetterFitness;
             copyTab(potentialBetterSolution, solution, size);
-            printf("\n== DEBUG == new best fitness : %d\n",fitness);
+            printf("\n== DEBUG == new best fitness : %f\n",fitness);
         }
     }
 
-    printf("{\"message\":\"finish\",\"best_fitness\":%d,\"best_solution\":[",fitness);
+    printf("{\"message\":\"finish\",\"best_fitness\":%f,\"best_solution\":[",fitness);
 
     for (int i = 0; i < size; ++i) {
         printf("%d", *(solution + i));
